+++ resources/libraries/get_sensors_info.rb
@@ -11,6 +11,11 @@ module RbManager
         sensors = search(:node, "role:#{s_type}#{s_type == 'intrusion-sensor' ? '' : ' AND -redborder_parent_id:*?'}").sort
         sensors_info[s_type] = {}
         sensors.each do |s|
+          # skip childs of proxy sensors
+          if s['redborder']['parent_id']
+            parent_sensor = search(:node, "sensor_id:#{s['redborder']['parent_id']}").first
+            next if parent_sensor && parent_sensor.to_s.include?('proxy')
+          end
           info = {}
           info['name'] = s.name
           info['ip'] = s['ipaddress']
@@ -27,6 +32,7 @@ module RbManager
 
             info['locations'][loc] = s['redborder'][loc]
           end
+
           sensors_info[s_type][s.name] = info
         end
       end